add token_cpi.rs
