+++ programs/drift_vaults/src/drift_cpi.rs
@@ -26,18 +26,6 @@ pub trait UpdateUserMarginTradingEnabledCPI {
     fn drift_update_user_margin_trading_enabled(&self, enabled: bool) -> Result<()>;
 }
 
-pub trait TokenTransferCPI {
-    fn token_transfer(&self, amount: u64) -> Result<()>;
-}
-
-pub trait MintTokensCPI {
-    fn mint(&self, vault_name: [u8; 32], vault_bump: u8, amount: u64) -> Result<()>;
-}
-
-pub trait BurnTokensCPI {
-    fn burn(&self, vault_name: [u8; 32], vault_bump: u8, amount: u64) -> Result<()>;
-}
-
 pub trait InitializeInsuranceFundStakeCPI {
     fn drift_initialize_insurance_fund_stake(&self, market_index: u16) -> Result<()>;
 }
