+++ programs/drift_vaults/src/instructions/protocol_withdraw.rs
@@ -9,8 +9,9 @@ use drift::state::user::User;
 use crate::constraints::{
     is_protocol_for_vault, is_user_for_vault, is_user_stats_for_vault, is_vault_protocol_for_vault,
 };
-use crate::drift_cpi::{TokenTransferCPI, WithdrawCPI};
+use crate::drift_cpi::WithdrawCPI;
 use crate::state::{Vault, VaultProtocol};
+use crate::token_cpi::TokenTransferCPI;
 use crate::{declare_vault_seeds, AccountMapProvider};
 
 pub fn protocol_withdraw<'c: 'info, 'info>(
